TodoApp/requirements.txt
fastapi == 0.115.5
uvicorn ==0.32.1
sqlalchemy ==2.0.36
mysql.connector ==2.2.9
mysqlclient
mysql-connector-python
emailvalidator
passlib
Bcrypt
python-jose
python-dotenv
pydantic_settings
openai
google-generativeai
google-genai